server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
google-genai
python-dotenv>=1.0.0
